LayoutTests/editing/selection/move-left-right.html

<head>
    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
    <style>
        div.test {
            -webkit-user-modify: read-write;
            padding: 4px;
            border: 1px dashed lightblue;
            margin: 4px 4px 4px 24px;
            outline: none;
            font-family: Lucida Grande;
            counter-increment: test-number;
        }
        div.test:before { content: counter(test-number); position: absolute; left: 8px; font-size: x-small; text-align: right; width: 20px; }
        div.test span { background-color: #def; }
        div.test img { width: 1em; height: 1em; background-color: lightgreen; }
        div.test img + img { background-color: lightblue; }
        div.test div { border: 1px dashed pink; padding: 3px; height: 2em; }
    </style>
    <script>
        function log(message)
        {
            document.getElementById("console").appendChild(document.createTextNode(message));
        }

        function caretCoordinates()
        {
            if (!window.textInputController)
                return { x: 0, y :0 };
            var caretRect = textInputController.firstRectForCharacterRange(textInputController.selectedRange()[0], 0);
            return { x: caretRect[0], y: caretRect[1] };
        }

        function positionsMovingInDirection(sel, direction)
        {
            var positions = [];
            while (true) {
                positions.push({ node: sel.anchorNode, offset: sel.anchorOffset, point: caretCoordinates() });
                sel.modify("move", direction, "character");
                if (positions[positions.length - 1].node == sel.anchorNode && positions[positions.length - 1].offset == sel.anchorOffset)
                    break;
            };
            return positions;
        }

        function fold(string)
        {
            var result = "";
            for (var i = 0; i < string.length; ++i) {
                var char = string.charCodeAt(i);
                if (char >= 0x05d0)
                    char -= 0x058f;
                else if (char == 10) {
                    result += "\\n";
                    continue;
                }
                result += String.fromCharCode(char);
            }
            return result;
        }

        function logPositions(positions)
        {
            for (var i = 0; i < positions.length; ++i) {
                if (i) {
                    if (positions[i].node != positions[i - 1].node)
                        log("]");
                    log(", ");
                }
                if (!i || positions[i].node != positions[i - 1].node)
                    log((positions[i].node instanceof Text ? '"' + fold(positions[i].node.data) + '"' : "<" + positions[i].node.tagName + ">") + "[");
                log(positions[i].offset);
            }
            log("]");
        }

        function checkReverseOrder(positions, reversePositions)
        {
            var mismatch = positions.length != reversePositions.length;
            while (!mismatch) {
                var pos = positions.pop();
                if (!pos)
                    break;
                var reversePos = reversePositions.shift();
                if (pos.node != reversePos.node || pos.offset != reversePos.offset)
                    mismatch = true;
            }

            if (mismatch)
                log("WARNING: Moving to the left did not visit the same positions in reverse order as moving to the right.\n");
        }

        function checkCoordinatesMovingRightDown(positions)
        {
            for (var i = 1; i < positions.length; ++i) {
                if (positions[i].point.y > positions[i - 1].point.y || positions[i].point.x < positions[i - 1].point.x && positions[i].point.y >= positions[i - 1].point.y)
                    log("WARNING: Moved in the wrong direction in step " + i + ": from (" + positions[i - 1].point.x + ", " + positions[i - 1].point.y + ") to (" + positions[i].point.x + ", " + positions[i].point.y + ").\n");
            }
        }

        function checkCoordinatesMovingLeftDown(positions)
        {
            for (var i = 1; i < positions.length; ++i) {
                if (positions[i].point.y > positions[i - 1].point.y || positions[i].point.x > positions[i - 1].point.x && positions[i].point.y >= positions[i - 1].point.y)
                    log("WARNING: Moved in the wrong direction in step " + i + ": from (" + positions[i - 1].point.x + ", " + positions[i - 1].point.y + ") to (" + positions[i].point.x + ", " + positions[i].point.y + ").\n");
            }
        }

        function checkCoordinatesMovingRightUp(positions)
        {
            for (var i = 1; i < positions.length; ++i) {
                if (positions[i].point.y < positions[i - 1].point.y || positions[i].point.x < positions[i - 1].point.x && positions[i].point.y <= positions[i - 1].point.y)
                    log("WARNING: Moved in the wrong direction in step " + i + ": from (" + positions[i - 1].point.x + ", " + positions[i - 1].point.y + ") to (" + positions[i].point.x + ", " + positions[i].point.y + ").\n");
            }
        }

        function checkCoordinatesMovingLeftUp(positions)
        {
            for (var i = 1; i < positions.length; ++i) {
                if (positions[i].point.y < positions[i - 1].point.y || positions[i].point.x > positions[i - 1].point.x && positions[i].point.y <= positions[i - 1].point.y)
                    log("WARNING: Moved in the wrong direction in step " + i + ": from (" + positions[i - 1].point.x + ", " + positions[i - 1].point.y + ") to (" + positions[i].point.x + ", " + positions[i].point.y + ").\n");
            }
        }

        onload = function()
        {
            if (!window.layoutTestController)
                return;

            layoutTestController.dumpAsText();

            var tests = document.getElementsByClassName("test");
            var sel = getSelection();
            for (var i = 0; i < tests.length; ++i) {
                var positionsMovingRight;
                var positionsMovingLeft;
    
                log("Test " + (i + 1) + ", LTR:\n  Moving right: ");
                sel.setPosition(tests[i], 0);
                positionsMovingRight = positionsMovingInDirection(sel, "right");
                logPositions(positionsMovingRight);
                log("\n");
                checkCoordinatesMovingRightDown(positionsMovingRight);

                log("  Moving left:  ");
                positionsMovingLeft = positionsMovingInDirection(sel, "left");
                logPositions(positionsMovingLeft);
                log("\n");
                checkCoordinatesMovingLeftUp(positionsMovingLeft);

                checkReverseOrder(positionsMovingLeft, positionsMovingRight);

                tests[i].style.direction = "rtl";

                log("Test " + (i + 1) + ", RTL:\n  Moving left: ");
                sel.setPosition(tests[i], 0);
                positionsMovingLeft = positionsMovingInDirection(sel, "left");
                logPositions(positionsMovingLeft);
                log("\n");
                checkCoordinatesMovingLeftDown(positionsMovingLeft);

                log("  Moving right:  ");
                positionsMovingRight = positionsMovingInDirection(sel, "right");
                logPositions(positionsMovingRight);
                log("\n");
                checkCoordinatesMovingRightUp(positionsMovingRight);

                checkReverseOrder(positionsMovingLeft, positionsMovingRight);

                tests[i].style.display = "none";
            }
        }

    </script>
</head>
<body>
    <div class="test">
        abc
    </div>

    <div class="test">
        ××‘×’
    </div>

    <div class="test"><br>abc
    </div>

    <div class="test"><br>××‘×’
    </div>

    <div class="test">
        abc××‘×’def
    </div>

    <div class="test">
        ××‘×’abc×“×”×•
    </div>

    <div class="test">
        abc××‘×’×“×”×•
    </div>

    <div class="test">
        ××‘×’abcdef
    </div>

    <div class="test">
        <span>abc</span>××‘×’def
    </div>

    <div class="test">
        <span>××‘×’</span>abc×“×”×•
    </div>

    <div class="test">abc××‘×’123×“×”×•def
    </div>

    <div class="test">abc××‘×’123
    </div>

    <div class="test">abc××‘×’123def
    </div>

    <div class="test">××‘×’123×“×”×•abc×–×—×˜456×™×›×œ
    </div>

    <div class="test" style="width: 120px;">
        before    ××—×¨×™ ×× ×¦×™×§×œ×•×¤×“×™×”
    </div>

    <div class="test" style="width: 120px;">
        ×œ×¤× ×™    after encyclopedia
    </div>

    <div class="test" contenteditable style="width: 120px;">
        before    ××—×¨×™ ×× ×¦×™×§×œ×•×¤×“×™×”
    </div>

    <div class="test" contenteditable style="width: 120px;">
        ×œ×¤× ×™    after encyclopedia
    </div>

    <div class="test" style="width: 100px;">
        This is ×™×•×ª×¨ ×¦×¨ ××™×ª×¨ the boxes. 
    </div>

    <div contenteditable class="test" style="width: 100px;">
        This is ×™×•×ª×¨ ×¦×¨ ××™×ª×¨ the boxes. 
    </div>

    <div class="test">
        Lorem
        <div></div>
        ipsum
    </div>

    <div class="test">
        ×¦×œ×—×ª
        <div></div>
        ××¦× ×¤×ª
    </div>

    <div class="test">
        abcdef××‘×’<img>×“×”×•
    </div>

    <div class="test">
        ××‘×’×“×”×•abc<img>def
    </div>

    <div class="test">
        abc<input>××‘×’<img><img>×“×”×•ghi
    </div>

    <div class="test">
        ××‘×’<input>abc<img><img>def×“×”×•
    </div>

    <div class="test">
        abc××‘×’<span>×“×”×•</span>
    </div>

    <div class="test">
        ××‘×’abc<span>def</span>
    </div>

    <div class="test">
        ab<span>c××‘×’def</span>
    </div>

    <div class="test">
        ××‘<span>×’abc×“×”×•</span>
    </div>

    <div class="test">
        abc<span>××‘×’def</span>
    </div>

    <div class="test">
        ××‘×’<span>abc×“×”×•</span>
    </div>

    <div class="test">
        abc×def
    </div>

    <div class="test">
        ××‘×’a×“×”×•
    </div>

    <div class="test">
        abc××‘×’<span>def</span>
    </div>

    <div class="test">
        ××‘×’abc<span>×“×”×•</span>
    </div>

    <div class="test">
        abc×<span>×‘×’def</span>
    </div>

    <div class="test">
        ××‘×’a<span>bc×“×”×•</span>
    </div>

    <div class="test" style="white-space: pre;">abc<!-- -->
<!-- -->def</div>

    <div class="test" style="white-space: pre;">???Ô?Õ<!-- -->
<!-- -->?Ò?Ó?¥</div>

    <pre id="console"></pre>
</body>
